update installation script

+++ install.sh
@@ -1,11 +1,6 @@
 #!/usr/bin/env sh
 
-if [ ! -d ~/.config ]; then
-    mkdir ~/.config
-fi
-
-if [ -e ~/.config/nvim ]; then
-    rm -rf ~/.config/nvim
-fi
+[ -d ~/.config ] || mkdir ~/.config
+[ -e ~/.config/nvim ] && rm -rf ~/.config/nvim
 
 ln -sf $(pwd) ~/.config/nvim
